move tests from retrieval_test into the main test file
