+++ generative_ai/embeddings/test_embeddings.py
@@ -14,6 +14,8 @@
 
 import backoff
 
+import code_retrieval_example
+import document_retrieval_example
 import generate_embeddings_with_lower_dimension
 
 from google.api_core.exceptions import ResourceExhausted
@@ -55,3 +57,33 @@ def test_multimodal_embedding_image_lower_dimension() -> None:
     assert len(embeddings.image_embedding) == 128
     assert embeddings.text_embedding is not None
     assert len(embeddings.text_embedding) == 128
+
+
+@backoff.on_exception(backoff.expo, ResourceExhausted, max_time=10)
+def test_text_embed_text() -> None:
+    texts = [
+        "banana bread?",
+        "banana muffin?",
+        "banana?",
+    ]
+    dimensionality = 256
+    embeddings = document_retrieval_example.embed_text(
+        texts, "RETRIEVAL_QUERY", dimensionality
+    )
+    assert [len(e) for e in embeddings] == [dimensionality or 768] * len(texts)
+
+
+@backoff.on_exception(backoff.expo, ResourceExhausted, max_time=10)
+def test_code_embed_text() -> None:
+    texts = [
+        "banana bread?",
+        "banana muffin?",
+        "banana?",
+    ]
+    dimensionality = 256
+    embeddings = code_retrieval_example.embed_text(
+        texts=texts,
+        task="CODE_RETRIEVAL_QUERY",
+        dimensionality=dimensionality,
+    )
+    assert [len(e) for e in embeddings] == [dimensionality or 768] * len(texts)
